requirements.txt
blinker==1.9.0
charset-normalizer==3.4.4
click==8.3.1
colorama==0.4.6
contourpy==1.3.3
cycler==0.12.1
Flask==3.1.2
Flask-Login==0.6.3
fonttools==4.61.1
gunicorn==23.0.0
imbalanced-learn==0.14.1
itsdangerous==2.2.0
Jinja2==3.1.6
joblib==1.5.3
kiwisolver==1.4.9
MarkupSafe==3.0.3
matplotlib==3.10.8
numpy==2.4.0
packaging==25.0
pandas==2.3.3
pillow==12.1.0
pyparsing==3.3.1
python-dateutil==2.9.0.post0
pytz==2025.2
reportlab==4.4.7
scikit-learn==1.8.0
scipy==1.16.3
six==1.17.0
sklearn-compat==0.1.5
threadpoolctl==3.6.0
tzdata==2025.3
Werkzeug==3.1.4
